requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
aiomysql
pydantic
pytest
pytest-asyncio